fixing warnings, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@9465 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/traci-server/TraCIServerAPI_MeMeDetector.cpp
@@ -83,7 +83,7 @@ TraCIServerAPI_MeMeDetector::processGet(TraCIServer &server, tcpip::Storage &inp
             break;
         case LAST_STEP_VEHICLE_NUMBER:
             tempMsg.writeUnsignedByte(TYPE_INTEGER);
-            tempMsg.writeInt((float) e3->getVehiclesWithin());
+            tempMsg.writeInt((int) e3->getVehiclesWithin());
             break;
         case LAST_STEP_MEAN_SPEED:
             tempMsg.writeUnsignedByte(TYPE_FLOAT);
@@ -106,7 +106,7 @@ TraCIServerAPI_MeMeDetector::processGet(TraCIServer &server, tcpip::Storage &inp
     server.writeStatusCmd(CMD_GET_MULTI_ENTRY_EXIT_DETECTOR_VARIABLE, RTYPE_OK, warning, outputStorage);
     // send response
     outputStorage.writeUnsignedByte(0); // command length -> extended
-    outputStorage.writeInt(1 + 4 + tempMsg.size());
+    outputStorage.writeInt(static_cast<int>(1 + 4 + tempMsg.size()));
     outputStorage.writeStorage(tempMsg);
     return true;
 }
